requirements.txt
neo4j
python-dotenv
Faker
uvicorn
fastapi